// src/store/modules/jt-cesium-vue/modules/toolbar/modules/nature/mutation-types.ts
export enum NatureMutationTypes {
  RESET_STATE = 'RESET_STATE',

  // nature
  SET_SHOW_SUN = 'SET_SHOW_SUN',
  SET_SHOW_MOON = 'SET_SHOW_MOON',
  SET_SHOW_SKY_ATMOSPHERE = 'SET_SHOW_SKY_ATMOSPHERE',
  SET_ENABLE_LIGHT = 'SET_ENABLE_LIGHT',
  SET_SHOW_SKY_BOX = 'SET_SHOW_SKY_BOX',
  SET_SHOW_SHADOW = 'SET_SHOW_SHADOW',
}