// src/store/modules/jt-cesium-vue/modules/toolbar/modules/nature/action-types.ts
export enum NatureActionTypes {
  RESET_STATE = 'RESET_STATE',

  // nature
  SWITCH_SHOW_SUN = 'SWITCH_SHOW_SUN',
  SWITCH_SHOW_MOON = 'SWITCH_SHOW_MOON',
  SWITCH_SHOW_SKY_ATMOSPHERE = 'SWITCH_SHOW_SKY_ATMOSPHERE',
  SWITCH_ENABLE_LIGHT = 'SWITCH_ENABLE_LIGHT',
  SWITCH_SHOW_SKY_BOX = 'SWITCH_SHOW_SKY_BOX',
  SWITCH_SHOW_SHADOW = 'SWITCH_SHOW_SHADOW',
}